[Feat] Modify Player Hang Controller
행 컨트롤러 에서 수행하던 기능을 레전드 컨트롤러가 담당하며 해당 스테이트에 진입했을때 수행할 기능을 정의합니다.
