+++ ItaCH_Smash_Legends/Assets/Script/Player_State/LegendHangState.cs
@@ -2,30 +2,29 @@ using UnityEngine;
 
 public class LegendHangState : LegendBaseState
 {
-    // LegnedController 완료시 리펙토링
-
-    private PlayerHangController _playerHangController;
-    private PlayerStatus _playerStatus;
     private EffectController _effectController;
 
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         base.OnStateEnter(animator, stateInfo, layerIndex);
-
         _effectController = animator.GetComponent<EffectController>();
-        _playerHangController = animator.GetComponent<PlayerHangController>();
-        _playerStatus = animator.GetComponent<PlayerStatus>();
 
-        _playerStatus.CurrentState = PlayerStatus.State.Hang;
+        legendController.OnFalling(animator).Forget();
+        _effectController.StartInvincibleFlashEffet(_effectController.FLASH_COUNT).Forget();
+    }
 
-        animator.SetBool(AnimationHash.Run, false);
-        _playerHangController.OnFalling().Forget();
+    override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    {
+        if (legendController.IsTriggered(ActionType.Jump))
+        {
+            animator.Play(AnimationHash.Jump);
+        }
     }
 
     override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
-        _playerHangController.OffConstraints();
+        legendController.OffConstraints();
         _effectController.StartInvincibleFlashEffet(_effectController.HANG_JUMP_FLASH_COUNT).Forget();
-        _playerHangController.TaskCancel.Cancel();
+        legendController.TaskCancel.Cancel();
     }
 }
